requirements.txt
openai
requests
flask
python-dotenv
pybybit